Change Grade table
Change the gradePoint attribute to a Decimal with 3 digits, 2 being decimals

+++ dbStructure.sql
@@ -333,7 +333,7 @@ CREATE TABLE StudentProgram (
 
 CREATE TABLE Grade (
     grade VARCHAR(3),
-    gradePoint INT,
+    gradePoint DECIMAL(3, 2),
     PRIMARY KEY (grade)
 )  ENGINE=INNODB;
